AWS mini app (upload)
Lambda Payload Limit
There is a hard limit of 6mb when it comes to AWS Lambda payload size. This means we cannot send more than 6mb of data to AWS Lambda in a single request.
Developers will typically run into this limit if their application was using AWS Lambda as the middle man between their client and their AWS S3 asset storage.
For example, if the application needs to upload an image or video, we will easily run over this limit when we send the image to our AWS Lambda function for upload.

How do we avoid this limit?
However, we can get around this limit by using an AWS S3 pre-signed url for uploads. This allows our AWS Lambda function to instead generate a pre-signed url and return it to our client, then our client can use that temporary pre-signed url to upload directly to our AWS S3 bucket.